requirements.txt
CherryPy==3.2.2
Flask==0.9
Jinja2==2.6
Markdoc==0.6.6
Markdown==2.2.1
Pygments==1.5
WebOb==1.2.3
Werkzeug==0.8.3
argparse==1.2.1
httplib2==0.8
pyzmq==13.0.2
six==1.3.0
stevedore==0.8
twilio==3.4.5
unittest2==0.5.1
virtualenv==1.9.1
virtualenv-clone==0.2.4
virtualenvwrapper==3.7.1
wsgiref==0.1.2
